Streamline IterateFeatures(Sorted) QRs
- Push down all methods / fields that belong to the more specific IterateFeaturesSorted implementation
- Improve en- / decryption of iteration handles to use a deterministic algorithm which enables cacheable handles and use hex representation which is url-friendly by itself (formerly was Base64)
